Fixed the problem with Utilz Class and added Vault in xDaily
I've found the problem with the utility class. The problem was that the
initialization part should be in the first line in the onEnable line.
Also I've added Vault to xDaily. There is a new VaultIntegration class
in xDaily for some next projects and the utility class could be apply
for next projects. Hopefully there are no errors or bugs in it.